Move futimes to writeContents
