+++ lib/dest/writeContents/index.js
@@ -2,11 +2,44 @@
 
 var fs = require('fs');
 
-var futimes = require('../../futimes');
 var writeDir = require('./writeDir');
 var writeStream = require('./writeStream');
 var writeBuffer = require('./writeBuffer');
 var writeSymbolicLink = require('./writeSymbolicLink');
+
+
+// http://stackoverflow.com/a/10589791/586382
+function validDate(date) {
+  return date instanceof Date && !isNaN(date.valueOf());
+}
+
+function futimes(fd, stat, cb) {
+  fs.fstat(fd, function(error, stats) {
+    if (error) {
+      return cb(error);
+    }
+
+    // `futimes` only works if we own the file
+    if (stats.uid !== process.getuid()) {
+      return cb();
+    }
+
+    // Given `mtime` is not valid, do nothing
+    var mtime = stat.mtime;
+    if (!validDate(mtime)) {
+      return cb();
+    }
+
+    // Given `atime` is not valid, assign a new one with current time
+    var atime = stat.atime;
+    if (!validDate(atime)) {
+      atime = new Date();
+    }
+
+    // Set file `atime` and `mtime` fields
+    fs.futimes(fd, atime, mtime, cb);
+  });
+}
 
 
 function writeContents(writePath, file, cb) {
